prevent keypresses from reaching the GUI
On the GX device, the gpio_keys based keyboard is the only
keyboard and is therefore used by the system. A key press
on the LCD therefore also navigates the menu, and can cause
settings to be changed or the device to be rebooted.

In a hilarious bug report, entering 5813 on a paygo device
would navigate to the reboot option and reboot the device.

To prevent this, we call the EVIOCGRAB ioctl on the input
device.