reduce amount of light needed to activate the backlight
The backlight would only activate at high levels of ambient
light, leaving the impression that it isn't working. An
analog level of 200 was empirically determined to be high enough
that the backlight's proximity to the sensor does not keep the
display on, but low enough that any sort of light brings the display
up.